+++ armonaut/auth/__init__.py
@@ -14,13 +14,17 @@
 
 from pyramid.authorization import ACLAuthorizationPolicy
 from pyramid_multiauth import MultiAuthenticationPolicy
-from armonaut.auth.services import database_login_factory, oauth_state_factory
-from armonaut.auth.interfaces import IUserService, IOAuthStateService
+from armonaut.auth.services import database_login_factory, oauth_factory
+from armonaut.auth.interfaces import IUserService, IOAuthService
 from armonaut.auth.policies import SessionAuthenticationPolicy
 
 
 REDIRECT_FIELD_NAME = 'next'
-OAUTH_SCOPES = ' '.join(['user', 'user:email', 'repo', 'repo:status', 'repo_deployment', 'read:org', 'write:repo_hook'])
+OAUTH_SCOPES = ' '.join([
+    'user', 'user:email', 'repo',
+    'repo:status', 'repo_deployment',
+    'read:org', 'write:repo_hook'
+])
 
 
 def _authenticate(userid, request):
@@ -46,7 +50,7 @@ def _user(request):
 
 def includeme(config):
     config.register_service_factory(database_login_factory, IUserService)
-    config.register_service_factory(oauth_state_factory, IOAuthStateService)
+    config.register_service_factory(oauth_factory, IOAuthService)
 
     config.set_authentication_policy(
         MultiAuthenticationPolicy([